requirements.txt
certifi==2024.7.4
charset-normalizer==3.2.0
Colr==0.9.1
idna==3.7
inquirerpy==0.3.4
markdown-it-py==3.0.0
mdurl==0.1.2
nest-asyncio==1.5.6
pfzy==0.3.4
prompt-toolkit==3.0.36
Pygments==2.15.1
pypresence==4.3.0
pywebview==5.3
PyYAML==6.0.1
requests==2.32.4
rich==13.5.1
urllib3==2.2.2
wcwidth==0.2.6
websocket-server==0.6.4
websockets==11.0.3

